+++ src/App/pages/HomePage.tsx
@@ -1,14 +1,14 @@
 import { useEffect, useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { TProduct } from '../../types/types';
+import { TProduct } from '../types/types';
 // import styles from './HomePage.module.scss';
-import { ProductList } from '../../components/ProductList';
-import CategoriesList from '../../components/CategoriesList';
-import { BrandsList } from '../../components/BrandList';
-import { getAllCategories, getAllProducts } from '../../services/productService';
-import countProducts from '../../utils/countProducts';
-import { SearchInput } from '../../components/SearchInput';
-import { SelectSortBy } from '../../components/SelectSortBy';
+import { ProductList } from '../components/ProductList';
+import CategoriesList from '../components/CategoriesList';
+import { BrandsList } from '../components/BrandList';
+import { getAllCategories, getAllProducts } from '../services/productService';
+import countProducts from '../utils/countProducts';
+import { SearchInput } from '../components/SearchInput';
+import { SelectSortBy } from '../components/SelectSortBy';
 
 function HomePage() {
   const [products, setProducts] = useState<TProduct[]>([]);
